% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_to_character.R
\name{formula_to_character}
\alias{formula_to_character}
\title{As character formula -- based on as.character.formula from formula.tools}
\usage{
formula_to_character(x, ...)
}
\arguments{
\item{x}{formula object}

\item{...}{further arguments pass to or from other methods}
}
\value{
A list of two items:  left hand side of formula, right hand side of formula
}
\description{
As character formula -- based on as.character.formula from formula.tools
}
\author{
Levi Moneyhun
}
