% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_coef_plot_notes.R
\name{make_coef_plot_notes}
\alias{make_coef_plot_notes}
\title{Rolling Event Study}
\usage{
make_coef_plot_notes(event_study, labels = NULL, source = "")
}
\arguments{
\item{event_study}{The output from rolling_event_study (with simple=FALSE).}

\item{labels}{A named character vector of labels corresponding to the variables in the regression (e.g., c("aaa_ret" = "AAA Inc.","bbb_ret" =  "B Co.")).}

\item{y_axis_unit}{unit to which y-axis limits should be expanded.}

\item{y_axis_breaks}{number of y-axis breaks to be attempted}

\item{x_axis_breaks}{number of x-axis breaks}
}
\value{
ggplot object
}
\description{
Rolling Event Study
}
\author{
Levi Moneyhun
}
