% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_resid_plot.R
\name{make_resid_plot}
\alias{make_resid_plot}
\title{Rolling Event Study}
\usage{
make_resid_plot(
  event_study,
  labels = NULL,
  y_axis_unit = 0.05,
  y_axis_breaks = 10,
  x_axis_interval = NULL,
  source = "",
  include_labs = TRUE
)
}
\arguments{
\item{event_study}{The output from rolling_event_study (with simple=FALSE).}

\item{labels}{A named character vector of labels corresponding to the variables in the regression (e.g., c("aaa_ret" = "AAA Inc.","bbb_ret" =  "B Co.")).}

\item{y_axis_unit}{unit to which y-axis limits should be expanded.}

\item{y_axis_breaks}{number of y-axis breaks to be attempted}

\item{x_axis_interval}{date interval for x-axis (e.g., "3 months")}
}
\value{
ggplot object
}
\description{
Rolling Event Study
}
\author{
Levi Moneyhun
}
