% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_event_study.R
\name{rolling_event_study}
\alias{rolling_event_study}
\title{Rolling Event Study}
\usage{
rolling_event_study(
  formula,
  data,
  pred_date_range,
  exclude_dates = NULL,
  id_dates = NULL,
  misrep_dates = lubridate::ymd(c()),
  disc_dates = lubridate::ymd(c()),
  rolling_window = NULL,
  roll_fixed = NULL,
  orth = NULL,
  p.val_thresh = 0.05,
  simple = FALSE
)
}
\arguments{
\item{formula}{A formula object describing the regression to be run.  If dummy dates are specified in 'id_dates', they will be added to this formula.}

\item{data}{A data.frame object containing the data for the regression and a 'date' variable.}

\item{pred_date_range}{A vector of the min and max date for the prediction range.}

\item{exclude_dates}{A vector of dates which should be excluded from the fitted regression models.}

\item{id_dates}{A vector of dates which should be 'dummied out' via indicator variables.}

\item{misrep_dates}{A vector of dates which should be flagged as misrepresentation dates in the output.  Does not affect the regression model.}

\item{disc_dates}{A vector of dates which should be flagged as disclosure dates in the output.  Does not affect the regression models.}

\item{rolling_window}{A number specifying the width of the rolling window.  E.g., a window of 50 means that the regression used to predict returns on day i will be fitted on data spanning from day i-50 to day i-1.}

\item{roll_fixed}{Either 'observations' or 'window'.  Where excluded dates fall within the rolling window for a given date, this specifies whether the number of observations should remain fixed--meaning the window must be extended--or whether the window should remain fixed--meaning the number of observations will be lower.}

\item{orth}{A formula object describing the orthogonalziation which should be applied (e.g., for y ~ x1 + x2, y in the final regression would be replaced by the residual from this regression.)}

\item{p.val_thresh}{Threshold for flagging residuals as significant.  5\% by default.}

\item{simple}{TRUE/FALSE indicating whether full output or just 'table' should be returned.  FALSE by default.}
}
\value{
If simple is FALSE, an list containing a table of event study results, which the regressions themselves as a column, and many other variables of interest relating to the event study.
}
\description{
Rolling Event Study
}
\author{
Levi Moneyhun
}
